Made some adjustments
