+++ index.html
@@ -1,20 +1,26 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/1.4.0/p5.js"></script>
+    <script src="./tiles/p5.js"></script>
     <meta charset="utf-8" />
     <title>Wave Function Collapse</title>
     <style>
       body {
-        background-color: #555;
+        background-repeat: no-repeat;
+        background: linear-gradient(to bottom right, #33c5ff 0%, #3366ff 100%);
+        background-attachment: fixed;
+      }
+      canvas{
+        display: block;
+        margin: auto;
       }
     </style>
   </head>
 
   <body>
     <main></main>
-    <script src="tile.js"></script>
-    <script src="cell.js"></script>
-    <script src="sketch.js"></script>
+    <script src="./board/tile.js"></script>
+    <script src="./board/cell.js"></script>
+    <script src="./board/sketch.js"></script>
   </body>
 </html>
